+++ libraries/arrays/mod.ts
@@ -11,7 +11,8 @@ export { reverseComparator } from './src/reverseComparator.ts';
 export { shuffle, shuffleInPlace } from './src/shuffle.ts';
 
 // Weighting functions
-export { accumulateWeights, getCumulativeWeights, toCumulativeValues } from './src/toCumulativeValues.ts';
+export { accumulateWeights } from './src/accumulateWeights.ts';
+export { getCumulativeWeights, toCumulativeValues } from './src/toCumulativeValues.ts';
 export { getWeightedIndex } from './src/getWeightedIndex.ts';
 export { pickWeightedIndex } from './src/pickWeightedIndex.ts';
 export { pickWeightedItem, toPickWeightedItem } from './src/pickWeightedItem.ts';